third_party/WebKit/LayoutTests/editing/execCommand/insert_list_in_noneditable_list_parent.html

<!doctype html>
 <script src="../../resources/testharness.js"></script>
 <script src="../../resources/testharnessreport.js"></script>
 <script src="../assert_selection.js"></script>
 <script>
test(() => assert_selection(
  [
  '<div>',
    '<dl>',
      '<div contenteditable>|</div>',
    '</dl>',
  '</div>'
  ],
  'InsertUnorderedList',
  [
  '<div>',
    '<dl>',
      '<div contenteditable>',
        '<ul><li>|<br></li></ul>',
      '</div>',
    '</dl>',
  '</div>'
  ]),
  '1 insertUnorderedList should work for editable div in dl.');

test(() => assert_selection(
  [
  '<div>',
    '<ul>',
      '<div contenteditable>|</div>',
    '</ul>',
  '</div>'
  ],
  'insertUnorderedList',
  [
  '<div>',
    '<ul>',
      '<div contenteditable>',
        '<ul><li>|<br></li></ul>',
      '</div>',
    '</ul>',
  '</div>'
  ]),
  '2 InsertUnorderedList should work for editable div in ul.');

test(() => assert_selection(
  [
  '<div>',
    '<ol>',
      '<div contenteditable>|</div>',
    '</ol>',
  '</div>'
  ],
  'insertUnorderedList',
  [
  '<div>',
    '<ol>',
      '<div contenteditable>',
        '<ul><li>|<br></li></ul>',
      '</div>',
    '</ol>',
  '</div>'
  ]),
  '3 insertUnorderedList should work for editable div in ol.');
</script>
